# api/openapi.yaml
openapi: 3.0.3
info:
  title: SukumadPro API
  version: 0.1.0
  description: |
    OpenAPI reference for the Gin HTTP API used by the web and desktop clients.

    This specification is intentionally aligned to the current router structure.
    Some response bodies are modeled generically where the implementation is still evolving.
servers:
  - url: /api/v1
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: List async jobs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Job list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
...
  /scheduler/jobs:
    get:
      tags: [Scheduler]
      operationId: listScheduledJobs
      summary: List scheduled jobs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scheduled job list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
    post:
      tags: [Scheduler]
      operationId: createScheduledJob
      summary: Create a scheduled job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledJobUpsertRequest'
      responses:
        '201':
          description: Scheduled job created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
  /scheduler/jobs/{id}:
    get:
      tags: [Scheduler]
      operationId: getScheduledJob
      summary: Get a scheduled job
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Scheduled job details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
    put:
      tags: [Scheduler]
      operationId: updateScheduledJob
      summary: Update a scheduled job
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledJobUpsertRequest'
      responses:
        '200':
          description: Scheduled job updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiTokenAuth:
      type: apiKey
      in: header
      name: X-API-Token
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Numeric resource identifier
      schema:
        type: integer
        format: int64
...
  schemas:
...
    GenericObjectResponse:
      type: object
      additionalProperties: true
...
    ScheduledJobUpsertRequest:
      type: object
      required:
        - code
        - name
        - jobCategory
        - jobType
        - scheduleType
        - scheduleExpr
        - timezone
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        jobCategory:
          type: string
          enum: [integration, maintenance]
        jobType:
          type: string
        scheduleType:
          type: string
          enum: [cron, interval]
        scheduleExpr:
          type: string
        timezone:
          type: string
        enabled:
          type: boolean
        allowConcurrentRuns:
          type: boolean
        config:
          type: object
          additionalProperties: true
    PaginationMeta:
      type: object
      additionalProperties: true
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    ListEnvelope:
      type: object
      additionalProperties: true
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...